+++ solr/core/src/test/org/apache/solr/cloud/TestStressInPlaceUpdates.java
@@ -262,8 +262,8 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
                     synchronized (model) {
                       DocInfo currInfo = model.get(id);
                       if (null != returnedVersion
-                          && (Math.abs(returnedVersion.longValue()) > Math.abs(currInfo.version))) {
-                        model.put(id, new DocInfo(returnedVersion.longValue(), 0, 0));
+                          && (Math.abs(returnedVersion) > Math.abs(currInfo.version))) {
+                        model.put(id, new DocInfo(returnedVersion, 0, 0));
                       }
                     }
 
@@ -356,8 +356,8 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
                     synchronized (model) {
                       DocInfo currInfo = model.get(id);
                       if (null != returnedVersion
-                          && (Math.abs(returnedVersion.longValue()) > Math.abs(currInfo.version))) {
-                        model.put(id, new DocInfo(returnedVersion.longValue(), nextVal1, nextVal2));
+                          && (Math.abs(returnedVersion) > Math.abs(currInfo.version))) {
+                        model.put(id, new DocInfo(returnedVersion, nextVal1, nextVal2));
                       }
                     }
                   }
@@ -434,9 +434,7 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
 
                     final Long foundVersion = (Long) actual.getFieldValue("_version_");
                     assertNotNull(msg, foundVersion);
-                    assertTrue(
-                        msg + "... solr doc has non-positive version???",
-                        0 < foundVersion.longValue());
+                    assertTrue(msg + "... solr doc has non-positive version???", 0 < foundVersion);
                     final Integer intVal = (Integer) actual.getFieldValue("val1_i_dvo");
                     assertNotNull(msg, intVal);
 
@@ -447,9 +445,9 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
                         msg
                             + " ...solr returned older version then model. "
                             + "should not be possible given the order of operations in writer threads",
-                        Math.abs(expected.version) <= foundVersion.longValue());
+                        Math.abs(expected.version) <= foundVersion);
 
-                    if (foundVersion.longValue() == expected.version) {
+                    if (foundVersion == expected.version) {
                       assertEquals(msg, expected.intFieldValue, intVal.intValue());
                       assertEquals(msg, expected.longFieldValue, longVal.longValue());
                     }
